### 2022-09-15 #4766 @editorialbot
[PRE REVIEW]: Spectral Connectivity: a python package for computing multitaper spectral estimates and frequency-domain brain connectivity measures on the CPU and GPU
**Submitting author:** @edeno (<a href="http://orcid.org/0000-0003-4606-087X"></a>)
**Repository:** https://github.com/Eden-Kramer-Lab/spectral_connectivity
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
