### 2022-09-15 #4766 @danielskatz
👋 @edeno - note that editorialbot has suggested some possible DOIs for some of your references, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
