+++ paper.bib
@@ -24,6 +24,7 @@
   year = {2007},
   pages = {163--166},
   publisher = {{IEEE}},
+  doi = {10.1109/ICDSP.2007.4288544},
   file = {/Users/edeno/Dropbox (Personal)/Papers/Digital Signal Processing, 2007 15th International Conference on-/2007/Digital Signal Processing, 2007 15th International Conference on-2007-Baccalá et al-Generalized partial directed coherence.pdf}
 }
 
