+++ paper.bib
@@ -51,6 +51,7 @@
   author = {Baccal{\'a}, Luiz Antonio and Sameshima, Koichi and Ballester, G and Valle, AC and Timo Iaria, C{\'e}sar},
   year = {1998},
   journal = {Resumos}
+  doi = {10.1007/s005290050005}
 }
 
 @article{CimenserTrackingbrainstates2011,
@@ -78,6 +79,7 @@
   eprint = {2201.11941},
   eprinttype = {arxiv},
   archiveprefix = {arXiv}
+  doi = {10.48550/arXiv.2201.11941}
 }
 
 @article{Delgado-SallentPhencyclidineinducedpsychosiscauses2021,
@@ -86,6 +88,7 @@
   year = {2021},
   journal = {bioRxiv},
   publisher = {{Cold Spring Harbor Laboratory}}
+  doi = {10.1101/2021.02.03.429582}
 }
 
 @article{DhamalaAnalyzinginformationflow2008,
@@ -131,6 +134,7 @@
   publisher = {{Elsevier}},
   issn = {0013-4694},
   isbn = {0013-4694},
+  doi = {10.1016/0013-4694(83)90235-3}
   file = {/Users/edeno/Dropbox (Personal)/Papers/Electroencephalography and Clinical Neurophysiology-/1983/Electroencephalography and clinical neurophysiology-1983-Gotman-Measurement of small time differences between EEG channels_2.pdf}
 }
 
@@ -156,6 +160,7 @@
   number = {3},
   pages = {203--210},
   issn = {0340-1200},
+  doi = {10.1007/BF00198091}
   file = {/Users/edeno/Dropbox (Personal)/Papers/Biological Cybernetics-/1991/Biological cybernetics-1991-Kaminski_Blinowska-A new method of the description of the information flow in the brain structures.pdf}
 }
 
@@ -184,6 +189,7 @@
   year = {2019},
   pages = {97--106},
   publisher = {{Springer}}
+  doi = {10.1007/978-3-662-58485-9_11}
 }
 
 @article{LachauxMeasuringphasesynchrony1999,
