+++ paper.bib
@@ -200,6 +200,7 @@
   volume = {8},
   number = {4},
   pages = {194--208},
+  doi = {10.1002/(sici)1097-0193(1999)8:4<194::aid-hbm4>3.0.co;2-c}
   file = {/Users/edeno/Dropbox (Personal)/Papers/Human Brain Mapping-/1999/Human brain mapping-1999-Lachaux et al-Measuring phase synchrony in brain signals.pdf}
 }
 
@@ -213,6 +214,7 @@
   pages = {9844--9858},
   publisher = {{Soc Neuroscience}},
   issn = {0270-6474}
+  doi = {10.1523/JNEUROSCI.0854-21.2021}
 }
 
 @article{NolteIdentifyingtruebrain2004,
@@ -279,6 +281,7 @@
   pages = {645709},
   publisher = {{Frontiers Media SA}},
   issn = {1662-5137}
+  doi = {10.3389/fnsys.2021.645709}
 }
 
 @article{Vinckimprovedindexphasesynchronization2011,
