+++ paper.bib
@@ -50,7 +50,7 @@
   title = {Studying the Interaction between Brain Structures via Directed Coherence and a Causality Test},
   author = {Baccal{\'a}, Luiz Antonio and Sameshima, Koichi and Ballester, G and Valle, AC and Timo Iaria, C{\'e}sar},
   year = {1998},
-  journal = {Resumos}
+  journal = {Resumos},
   doi = {10.1007/s005290050005}
 }
 
@@ -78,7 +78,7 @@
   journal = {arXiv preprint arXiv:2201.11941},
   eprint = {2201.11941},
   eprinttype = {arxiv},
-  archiveprefix = {arXiv}
+  archiveprefix = {arXiv},
   doi = {10.48550/arXiv.2201.11941}
 }
 
@@ -87,7 +87,7 @@
   author = {{Delgado-Sallent}, Cristina and Nebot, Pau and Gener, Thomas and Timplalexi, Melina and Fath, Amanda B. and Puig, M. Victoria},
   year = {2021},
   journal = {bioRxiv},
-  publisher = {{Cold Spring Harbor Laboratory}}
+  publisher = {{Cold Spring Harbor Laboratory}},
   doi = {10.1101/2021.02.03.429582}
 }
 
@@ -134,7 +134,7 @@
   publisher = {{Elsevier}},
   issn = {0013-4694},
   isbn = {0013-4694},
-  doi = {10.1016/0013-4694(83)90235-3}
+  doi = {10.1016/0013-4694(83)90235-3},
   file = {/Users/edeno/Dropbox (Personal)/Papers/Electroencephalography and Clinical Neurophysiology-/1983/Electroencephalography and clinical neurophysiology-1983-Gotman-Measurement of small time differences between EEG channels_2.pdf}
 }
 
@@ -160,7 +160,7 @@
   number = {3},
   pages = {203--210},
   issn = {0340-1200},
-  doi = {10.1007/BF00198091}
+  doi = {10.1007/BF00198091},
   file = {/Users/edeno/Dropbox (Personal)/Papers/Biological Cybernetics-/1991/Biological cybernetics-1991-Kaminski_Blinowska-A new method of the description of the information flow in the brain structures.pdf}
 }
 
@@ -188,7 +188,7 @@
   author = {K{\"u}hnert, Christian and Frey, Christian and Seyboldt, Ruben},
   year = {2019},
   pages = {97--106},
-  publisher = {{Springer}}
+  publisher = {{Springer}},
   doi = {10.1007/978-3-662-58485-9_11}
 }
 
@@ -200,7 +200,7 @@
   volume = {8},
   number = {4},
   pages = {194--208},
-  doi = {10.1002/(sici)1097-0193(1999)8:4<194::aid-hbm4>3.0.co;2-c}
+  doi = {10.1002/(sici)1097-0193(1999)8:4<194::aid-hbm4>3.0.co;2-c},
   file = {/Users/edeno/Dropbox (Personal)/Papers/Human Brain Mapping-/1999/Human brain mapping-1999-Lachaux et al-Measuring phase synchrony in brain signals.pdf}
 }
 
@@ -214,7 +214,7 @@
   pages = {9844--9858},
   publisher = {{Soc Neuroscience}},
   issn = {0270-6474}
-  doi = {10.1523/JNEUROSCI.0854-21.2021}
+  doi = {10.1523/JNEUROSCI.0854-21.2021},
 }
 
 @article{NolteIdentifyingtruebrain2004,
@@ -280,7 +280,7 @@
   volume = {15},
   pages = {645709},
   publisher = {{Frontiers Media SA}},
-  issn = {1662-5137}
+  issn = {1662-5137},
   doi = {10.3389/fnsys.2021.645709}
 }
 
