+++ paper.bib
@@ -213,8 +213,8 @@
   number = {47},
   pages = {9844--9858},
   publisher = {{Soc Neuroscience}},
-  issn = {0270-6474}
-  doi = {10.1523/JNEUROSCI.0854-21.2021},
+  issn = {0270-6474},
+  doi = {10.1523/JNEUROSCI.0854-21.2021}
 }
 
 @article{NolteIdentifyingtruebrain2004,
