### 2022-09-21 a205638
Fix xarray interface and bring back tests
