### 2022-09-26 #4766 @emdupre
👋 Hi @edeno, and thanks for your submission to JOSS ! I see that you have already suggested potential reviewers in thread ; I'll look through these suggestions and send out invitations soon.
