+++ spectral_connectivity/transforms.py
@@ -7,15 +7,17 @@ from scipy.linalg import eigvals_banded
 
 logger = getLogger(__name__)
 
-if os.environ.get('SPECTRAL_CONNECTIVITY_ENABLE_GPU') == 'true':
+if os.environ.get("SPECTRAL_CONNECTIVITY_ENABLE_GPU") == "true":
     try:
-        logger.info('Using GPU for spectral_connectivity...')
+        logger.info("Using GPU for spectral_connectivity...")
         import cupy as xp
         from cupy.linalg import lstsq
         from cupyx.scipy.fft import fft, fftfreq, ifft, next_fast_len
     except ImportError:
-        print('Cupy not installed. Cupy is needed to use GPU for '
-              'spectral_connectivity.')
+        print(
+            "Cupy not installed. Cupy is needed to use GPU for "
+            "spectral_connectivity."
+        )
         import numpy as xp
         from scipy.fftpack import fft, fftfreq, ifft, next_fast_len
         from scipy.linalg import lstsq
@@ -26,7 +28,7 @@ else:
 
 
 class Multitaper(object):
-    '''Transform time-domain signal(s) to the frequency domain by using
+    """Transform time-domain signal(s) to the frequency domain by using
     multiple tapering windows.
 
     Attributes
@@ -65,15 +67,24 @@ class Multitaper(object):
     is_low_bias : bool, optional
         If `True`, excludes tapers with eigenvalues < 0.9
 
-    '''
+    """
 
-    def __init__(self, time_series, sampling_frequency=1000,
-                 time_halfbandwidth_product=3,
-                 detrend_type='constant', time_window_duration=None,
-                 time_window_step=None, n_tapers=None,  tapers=None,
-                 start_time=0, n_fft_samples=None,
-                 n_time_samples_per_window=None,
-                 n_time_samples_per_step=None, is_low_bias=True):
+    def __init__(
+        self,
+        time_series,
+        sampling_frequency=1000,
+        time_halfbandwidth_product=3,
+        detrend_type="constant",
+        time_window_duration=None,
+        time_window_step=None,
+        n_tapers=None,
+        tapers=None,
+        start_time=0,
+        n_fft_samples=None,
+        n_time_samples_per_window=None,
+        n_time_samples_per_step=None,
+        is_low_bias=True,
+    ):
 
         self.time_series = xp.asarray(time_series)
         self.sampling_frequency = sampling_frequency
@@ -91,74 +102,81 @@ class Multitaper(object):
 
     def __repr__(self):
         return (
-            'Multitaper('
-            'sampling_frequency={0.sampling_frequency!r}, '
-            'time_halfbandwidth_product={0.time_halfbandwidth_product!r},\n'
-            '           time_window_duration={0.time_window_duration!r}, '
-            'time_window_step={0.time_window_step!r},\n'
-            '           detrend_type={0.detrend_type!r}, '
-            'start_time={0.start_time}, '
-            'n_tapers={0.n_tapers}'
-            ')'.format(self))
+            "Multitaper("
+            "sampling_frequency={0.sampling_frequency!r}, "
+            "time_halfbandwidth_product={0.time_halfbandwidth_product!r},\n"
+            "           time_window_duration={0.time_window_duration!r}, "
+            "time_window_step={0.time_window_step!r},\n"
+            "           detrend_type={0.detrend_type!r}, "
+            "start_time={0.start_time}, "
+            "n_tapers={0.n_tapers}"
+            ")".format(self)
+        )
 
     @property
     def tapers(self):
-        '''
+        """
 
         Returns
         -------
         tapers : array_like, shape (n_time_samples_per_window, n_tapers)
 
-        '''
+        """
         if self._tapers is None:
             self._tapers = _make_tapers(
[…]
