+++ spectral_connectivity/connectivity.py
@@ -6,14 +6,18 @@ from itertools import combinations
 import numpy as np
 from scipy.ndimage import label
 from scipy.stats.mstats import linregress
-from spectral_connectivity.minimum_phase_decomposition import \
-    minimum_phase_decomposition
-from spectral_connectivity.statistics import (adjust_for_multiple_comparisons,
-                                              coherence_bias,
-                                              fisher_z_transform,
-                                              get_normal_distribution_p_values)
 
-if os.environ.get('SPEC_CON_ENABLE_GPU') == 'true':
+from spectral_connectivity.minimum_phase_decomposition import (
+    minimum_phase_decomposition,
+)
+from spectral_connectivity.statistics import (
+    adjust_for_multiple_comparisons,
+    coherence_bias,
+    fisher_z_transform,
+    get_normal_distribution_p_values,
+)
+
+if os.environ.get("SPEC_CON_ENABLE_GPU") == "true":
     try:
         import cupy as xp
         from cupyx.scipy.fft import ifft
@@ -28,18 +32,19 @@ else:
     from scipy.sparse.linalg import svds
 
 EXPECTATION = {
-    'time': partial(xp.mean, axis=0),
-    'trials': partial(xp.mean, axis=1),
-    'tapers': partial(xp.mean, axis=2),
-    'time_trials': partial(xp.mean, axis=(0, 1)),
-    'time_tapers': partial(xp.mean, axis=(0, 2)),
-    'trials_tapers': partial(xp.mean, axis=(1, 2)),
-    'time_trials_tapers': partial(xp.mean, axis=(1, 2, 3)),
+    "time": partial(xp.mean, axis=0),
+    "trials": partial(xp.mean, axis=1),
+    "tapers": partial(xp.mean, axis=2),
+    "time_trials": partial(xp.mean, axis=(0, 1)),
+    "time_tapers": partial(xp.mean, axis=(0, 2)),
+    "trials_tapers": partial(xp.mean, axis=(1, 2)),
+    "time_trials_tapers": partial(xp.mean, axis=(1, 2, 3)),
 }
 
 
 def asnumpy(connectivity_measure):
     """If cupy array, then return as numpy. Else return numpy array"""
+
     @wraps(connectivity_measure)
     def wrapper(*args, **kwargs):
         measure = connectivity_measure(*args, **kwargs)
@@ -50,11 +55,13 @@ def asnumpy(connectivity_measure):
                 return measure
         else:
             return None
+
     return wrapper
 
 
 def non_negative_frequencies(axis):
-    '''Decorator that removes the negative frequencies.'''
+    """Decorator that removes the negative frequencies."""
+
     def decorator(connectivity_measure):
         @wraps(connectivity_measure)
         def wrapper(*args, **kwargs):
@@ -65,7 +72,9 @@ def non_negative_frequencies(axis):
                 return xp.take(measure, indices=non_neg_index, axis=axis)
             else:
                 return None
+
         return wrapper
+
     return decorator
 
 
@@ -77,7 +86,7 @@ def nonsorted_unique(x):
 
 
 class Connectivity:
-    '''Computes brain connectivity measures based on the cross spectral
+    """Computes brain connectivity measures based on the cross spectral
     matrix.
 
     Spectral granger methods that require estimation of transfer function
@@ -132,11 +141,17 @@ class Connectivity:
            information flow in brain networks with noxparametric Granger
            causality. NeuroImage 41, 354-362.
 
-    '''
+    """
 
-    def __init__(self, fourier_coefficients,
-                 expectation_type='trials_tapers', frequencies=None,
-                 time=None, blocks=None, dtype=xp.complex128):
+    def __init__(
+        self,
+        fourier_coefficients,
[…]
