### 2022-10-04 a554f23
Add notebook integration test
