### 2022-10-06 #4766 @emdupre
👋 Hi @jasmainak @sappelhoff @ajquinn @rly !

Would you be willing to review _Spectral Connectivity: a python package for computing multitaper spectral estimates and frequency-domain brain connectivity measures on the CPU and GPU_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/Eden-Kramer-Lab/spectral_connectivity) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.04766/joss.04766/10.21105.joss.04766.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you !
