### 2022-10-07 #4766 @edeno
Hello @emdupre ,

Thank you for the requests and editing this review.

+ I believe the proper spelling has a comma in between California and San Francisco. Or at least this is how it is written in the brand guidelines for the university: https://identity.ucsf.edu/brand-guide/using-ucsf-name. I have no particular attachment to this, so I'm happy to change it. I will double check the other affiliations as well.

+ I will add the contributing guidelines right away.

Thank you again.
