### 2022-10-10 #4766 @emdupre
Thanks for letting us know, @rly, @sappelhoff, @EtienneCmb !

@sappelhoff and @EtienneCmb, I'll add you both with @AJQuinn as reviewers on the submission now.

@edeno, thanks for double-checking those details ! I've re-checked the rendering and I don't think the comma is splitting the affiliation in this format, so we should be OK as-is.

I'll go ahead and close this discussion to open the official review issue 🚀
