### 2022-10-12 Eden-Kramer-Lab/spectral_connectivity#32 @sappelhoff
[JOSS] Add `.mailmap` file
While checking the authors of this software for JOSS (https://github.com/openjournals/joss-reviews/issues/4840), I used the following command: `git shortlog --email --summary`, and it gave me the following list:

```
     3  Emily Stephen <emilyps14@gmail.com>
     2  Emily Stephen <emilyps@mit.edu>
   121  Eric Denovellis <edeno@bu.edu>
    22  Eric Denovellis <edeno@users.noreply.github.com>
     5  Eric Denovellis <edenovellis@gmail.com>
     5  Max Myroshnychenko <mmyros@gmail.com>
    16  Mehrad <59752677+MehradSm@users.noreply.github.com>
     1  dizcza <dizcza@gmail.com>
    26  edeno <edeno@bu.edu>
     1  myroshnychenkm2 <mmyros@gmail.com>
```

As you can see, several authors appear multiple times under different aliases and emails. This is something you can solve with a `.mailmap` file, see: https://git-scm.com/docs/gitmailmap

For an example, see: https://github.com/mne-tools/mne-bids/blob/main/.mailmap
