### 2022-10-12 Eden-Kramer-Lab/spectral_connectivity#34 @sappelhoff
[JOSS] Add badges to README
It'd be very convenient to get the following "badges" from your README:

- [travis status](https://github.com/Eden-Kramer-Lab/spectral_connectivity/blob/master/.travis.yml), given that this CI is still running as expected. If it doesn't now would be a good time to remove or fix it
- [PyPI status](https://pypi.org/project/spectral-connectivity/)
- Conda status (I can't find your "edeno" channel and the package right now)
- [readthedocs](https://spectral-connectivity.readthedocs.io/en/latest/)

Furthermore, it'd be nice to have a badge for test coverage, for example using https://about.codecov.io/
See this repo for an example: https://github.com/bids-standard/pybv/blob/4af6bbab86e6151a3a64d759d4eca4a15e9b253a/.github/workflows/python_tests.yml#L81-L85

xref: https://github.com/openjournals/joss-reviews/issues/4840
