### 2022-10-12 Eden-Kramer-Lab/spectral_connectivity#35 @sappelhoff
[JOSS] run CI also on PRs
I noticed that your CI tests don't seem to run on PRs (see e.g., #33).

That might be because you defined too few "triggers": https://github.com/Eden-Kramer-Lab/spectral_connectivity/blob/d4ade18825c63050e3297909e64232c547178c5c/.github/workflows/PR-test.yml#L3

for an example on how to define triggers that also work on pull requests, see: https://github.com/sappelhoff/pyprep/blob/c6b268b90e08af2d4740496414cad649344345fa/.github/workflows/python_tests.yml#L3-L12

I'd strongly recommend running CIs for PRs as well.

xref: https://github.com/openjournals/joss-reviews/issues/4840
