### 2022-10-12 Eden-Kramer-Lab/spectral_connectivity#36 @sappelhoff
[JOSS] Extend CONTRIBUTING.md and add CODE_OF_CONDUCT.md
I recommend that you extend your contributor guidelines, they could additionally contain:

- how to build docs
- how to make a release (or at least, what the processes are -- especially given that you use both pypi and conda)

Next to that I recommend that you include a code of conduct file, for example following: https://www.contributor-covenant.org/

xref: https://github.com/openjournals/joss-reviews/issues/4840
