### 2022-10-12 Eden-Kramer-Lab/spectral_connectivity#37 @sappelhoff
[JOSS] Consider moving to conda-forge?
I am curious: Why do you ship `spectral_connectivity` on a personal conda channel, instead of on [conda-forge](https://conda-forge.org/)?

Shipping on conda-forge would have the advantage of re-using the huge community structure, and making your release process for conda more transparent, as well as users not having to declare your personal channel as a source in the environments.

xref: https://github.com/openjournals/joss-reviews/issues/4840
