### 2022-10-12 Eden-Kramer-Lab/spectral_connectivity#38 @sappelhoff
[JOSS] Docs API section does not show function signatures
See: https://spectral-connectivity.readthedocs.io/en/latest/api.html

Given that you have nice numpydoc style docstrings for your functions and classes, it'd be nice to also be able to see them in your automatically generated docs.

This repo contains a setup similar to yours (functions and classes documented via numpydoc, built via sphinx and readthedocs), so you may find it useful to debug your own docs: https://pyprep.readthedocs.io/en/latest/api.html

xref: https://github.com/openjournals/joss-reviews/issues/4840
