### 2022-10-12 Eden-Kramer-Lab/spectral_connectivity#39 @sappelhoff
[JOSS] remove unneeded CI steps?
Is the `draft-pdf.yml` CI workflow still needed now that the JOSS review is under way and your paper builds successfully (as can be checked in https://github.com/openjournals/joss-reviews/issues/4840)? If not, I recommend removing it.

https://github.com/Eden-Kramer-Lab/spectral_connectivity/blob/d4ade18825c63050e3297909e64232c547178c5c/.github/workflows/draft-pdf.yml#L1
