### 2022-10-12 Eden-Kramer-Lab/spectral_connectivity#40 @sappelhoff
[JOSS] reduce superflous files?
Removing your `setup.cfg` file would remove one file (less clutter in the top level of the repo) while not being too much of a drawback, because all that's specified is:

https://github.com/Eden-Kramer-Lab/spectral_connectivity/blob/d4ade18825c63050e3297909e64232c547178c5c/setup.cfg#L1-L2

if you really want that alias, you could also define a step like that in your [Makefile](https://github.com/Eden-Kramer-Lab/spectral_connectivity/blob/d4ade18825c63050e3297909e64232c547178c5c/Makefile#L1)

xref: https://github.com/openjournals/joss-reviews/issues/4840
