### 2022-10-12 Eden-Kramer-Lab/spectral_connectivity#41 @sappelhoff
[JOSS] double check CITATION.cff and prepare transparent "author" requirements
See for example:

https://github.com/Eden-Kramer-Lab/spectral_connectivity/blob/d4ade18825c63050e3297909e64232c547178c5c/CITATION.cff#L25

I think you can be more specific in what license you use: According to https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md#license, and your LICENSE file, it should be: `GPL-3.0`

Apart from that I see that you have had one more person who committed code to this repo (`dizcza <dizcza@gmail.com>`) but they are not in the CITATION.cff file (and not in the JOSS publication). If this is a conscious decision, it may be good to clarify in your CONTRIBUTORS.md file, at what point contributors are considered co-authors of the software.

xref: https://github.com/openjournals/joss-reviews/issues/4840
