### 2022-10-12 58a3cc3
minor changes README
