+++ README.md
@@ -13,7 +13,7 @@
 
 ### What is spectral_connectivity? ###
 
-`spectral_connectivity` is a python software package that computes multitaper spectral estimates and frequency-domain brain connectivity measures such as coherence, spectral granger causality, and the phase lag index using the multitaper Fourier transform. Although there are other python packages that do this (see [nitime](https://github.com/nipy/nitime) and [MNE-Python](https://github.com/mne-tools/mne-python)), spectral has several differences:
+`spectral_connectivity` is a Python software package that computes multitaper spectral estimates and frequency-domain brain connectivity measures such as coherence, spectral granger causality, and the phase lag index using the multitaper Fourier transform. Although there are other Python packages that do this (see [nitime](https://github.com/nipy/nitime) and [MNE-Python](https://github.com/mne-tools/mne-python)), `spectral_connectivity` has several differences:
 
 + it is designed to handle multiple time series at once
 + it caches frequently computed quantities such as the cross-spectral matrix and minimum-phase-decomposition, so that connectivity measures that use the same processing steps can be more quickly computed.
@@ -26,7 +26,10 @@ efficiently summarize brain-area level coherences from multielectrode recordings
 
 ### Tutorials ###
 
-See the notebooks ([\#1](examples/Tutorial_On_Simulated_Examples.ipynb), [\#2](examples/Tutorial_Using_Paper_Examples.ipynb)) for more information on how to use the package.
+See the following notebooks for more information on how to use the package:
+
++ [Tutorial on simulated examples](examples/Tutorial_On_Simulated_Examples.ipynb)
++ [Tutorial using paper examples](examples/Tutorial_Using_Paper_Examples.ipynb)
 
 ### Usage Example ###
 
@@ -95,13 +98,13 @@ See [environment.yml](environment.yml) for the most current list of dependencies
 
 ### Installation ###
 
-```python
+```bash
 pip install spectral_connectivity
 ```
 
 or
 
-```python
+```bash
 conda install -c edeno spectral_connectivity
 ```
 
