### 2022-10-12 Eden-Kramer-Lab/spectral_connectivity#43 @sappelhoff
[JOSS] make better use of the metadata options for PyPI
currently the pypi metadata is very bare-bones: https://pypi.org/project/spectral-connectivity/

I recommend that you add as much metadata as you can to your `setup.py` file: https://github.com/Eden-Kramer-Lab/spectral_connectivity/blob/master/setup.py

For an example, see: https://github.com/autoreject/autoreject/blob/master/setup.py

xref: https://github.com/openjournals/joss-reviews/issues/4840
