### 2022-10-14 712b2ad
Remove setup.cfg and JOSS draft action
Addresses issue #40  and #39
