### 2022-10-14 052fa39
Run CI on PRs
Addresses issue #35
