+++ .github/workflows/PR-test.yml
@@ -1,6 +1,10 @@
 name: PR Test
 
-on: push
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
 
 jobs:
   run-tests:
