### 2022-10-14 39ef7db
Remove travis CI yaml because we use github actions
Addresses issue #34
