+++ .travis.yml
@@ -1,20 +1,0 @@
-language: python
-python:
-    - '3.5'
-notifications:
-  email: false
-install:
-  - sudo apt-get update
-  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
-  - bash miniconda.sh -b -p $HOME/miniconda
-  - export PATH="$HOME/miniconda/bin:$PATH"
-  - hash -r
-  - conda config --set always_yes yes --set changeps1 no
-  - conda update -q conda
-  - conda info -a
-  - conda env create -f environment.yml
-  - source activate spectral_connectivity
-  - python setup.py develop
-script: pytest --cov=spectral_connectivity tests/
-after_success:
-    - coveralls
