### 2022-10-14 c3e8218
Add coveralls for code coverage
