### 2022-10-14 22281c7
Fix mapping on mailmap
Addresses issue #32
