### 2022-10-14 b4073de
Fix license type
Addresses issue #41
