+++ CITATION.cff
@@ -22,4 +22,4 @@ title: Eden-Kramer-Lab/spectral_connectivity
 version: v1.0.3
 date-released: 2022-09-14
 repository-code: https://github.com/Eden-Kramer-Lab/spectral_connectivity
-license: other-open
+license: GPL-3.0
