### 2022-10-14 Eden-Kramer-Lab/spectral_connectivity#41 @edeno
dizcza had only minimally contributed and had declined participation in the paper. However, I will add clarification in the CONTRIBUTORS.md
