### 2022-10-14 f987066
Add more metadata
Addresses issue #43
