+++ setup.py
@@ -23,8 +23,31 @@ setup(
     author="Eric Denovellis",
     author_email="edeno@bu.edu",
     url="https://github.com/Eden-Kramer-Lab/spectral_connectivity",
+    long_description=open("README.md").read(),
+    long_description_content_type="text/x-rst",
+    classifiers=[
+        "Intended Audience :: Science/Research",
+        "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
+        "Programming Language :: Python",
+        "Topic :: Scientific/Engineering",
+        "Operating System :: Microsoft :: Windows",
+        "Operating System :: POSIX",
+        "Operating System :: Unix",
+        "Operating System :: MacOS",
+        "Programming Language :: Python :: 3",
+    ],
+    platforms="any",
+    keywords=(
+        "python neuroscience electrophysiology "
+        "multitaper spectrogram frequency-domain"
+    ),
     python_requires=">=3",
     packages=find_packages(),
     install_requires=INSTALL_REQUIRES,
     tests_require=TESTS_REQUIRE,
+    project_urls={
+        "Documentation": "https://spectral-connectivity.readthedocs.io/en/latest/",
+        "Bug Reports": "https://github.com/Eden-Kramer-Lab/spectral_connectivity/issues",
+        "Source": "https://github.com/Eden-Kramer-Lab/spectral_connectivity",
+    },
 )
