+++ paper.bib
@@ -338,3 +338,30 @@
   journal = {GitHub repository},
   url = {https://github.com/nipy/nitime/}
 }
+
+
+@book{MullerPythonNeuroscience2015,
+  title = {Python in {{Neuroscience}}},
+  editor = {Muller, Eilif and Bednar, James A. and Diesmann, Markus and Gewaltig, Marc-Oliver and Hines, Michael and Davison, Andrew P.},
+  year = {2015},
+  series = {Frontiers {{Research Topics}}},
+  publisher = {{Frontiers Media SA}},
+  doi = {10.3389/978-2-88919-608-1},
+  isbn = {978-2-88919-608-1},
+  langid = {english},
+  file = {/Users/edeno/Zotero/storage/G7743XGF/Muller et al. - 2015 - Python in Neuroscience.PDF}
+}
+
+@techreport{SchlaflyPythonpracticingneuroscientist2020,
+  type = {Preprint},
+  title = {Python for the Practicing Neuroscientist: An Online Educational Resource},
+  shorttitle = {Python for the Practicing Neuroscientist},
+  author = {Schlafly, Emily and Cheung, Anthea and Michalka, Samantha W and Lipton, Paul A and Kochlacs, Caroline Moore and Bohland, Jason and Eden, Uri T and Kramer, Mark},
+  year = {2020},
+  month = jul,
+  institution = {{Preprints}},
+  doi = {10.22541/au.159363438.81020330},
+  abstract = {As neuronal data accumulates worldwide, accessible - yet rigorous - resources to develop hands-on experience with modern data analysis techniques are required. We present here an online educational resource for neural data analysis (https://markkramer.github.io/Case-Studies-Python). To reach the biologists, psychologists, and clinicians collecting neuronal data, we assume only a basic mathematics background, common to those trained in biological sciences. Through an interdisciplinary case-study approach, we use real-world data to motivate the study of modern quantitative analysis methods in Python. A modular format provides multiple coherent learning paths through the material, and thereby allows personalized learning for individuals with varying quantitative backgrounds and research interests, and flexible curation of material for redeployment in other curricula. Developed using Jupyter notebooks, the material supports fully interactive environments in most web browsers, and hosted on GitHub, the material is freely available for reuse, modification, and further development by the community.},
+  langid = {english},
+  file = {/Users/edeno/Zotero/storage/48VNSJCT/Schlafly et al. - 2020 - Python for the practicing neuroscientist an onlin.pdf}
+}
