+++ paper.md
@@ -45,7 +45,7 @@ In neuroscience, characterizing the oscillatory dynamics of the brain is critica
 
 # Statement of Need
 
-`spectral_connectivity` is a Python software package that computes multitaper spectral estimates and frequency-domain brain connectivity measures. The programming language Python is increasingly being used in the neurosciences, but the two main packages for spectral analysis in Python, `nitime` [@Rokem2020] and `mne-python` [@GramfortEtAl2013a], have issues that make them more difficult to use in many situations. For example, `nitime` implements several estimators of the power spectrum, but lacks spectrograms and windowed spectral estimators. `mne-python` is a much larger package designed as a full-featured analysis library for EEG and MEG data, and using the spectral analysis functions requires representing data using its ecosystem. Users of other data modalities, such as non-human electrophysiology data, may thus find much of `mne-python` inappropriate or unsuited for their analyses. This is an important problem because the non-human neurosciences are undergoing a period of great technological development; more and more signals are being collected simultaneously, and the duration of these signals is becoming longer as chronic recordings become possible. This rapid increase in the size and duration of datasets demands a lightweight, fast, and efficient spectral estimation package. `spectral_connectivity` is designed to handle multiple time series and can exploit GPUs for faster and more efficient computation. In addition, it can block compute important quantities such as the cross-spectral matrix in order to reduce memory burdens caused by large datasets. `spectral_connectivity` is also designed to be a lightweight package that has a simple user interface and can be easily be incorporated with other packages. Finally, `spectral_connectivity` also implements several connectivity measures that have not previously been implemented in Python such as the non-parametric version of the spectral granger causality and canonical coherence.
+`spectral_connectivity` is a Python software package that computes multitaper spectral estimates and frequency-domain brain connectivity measures. The programming language Python is increasingly being used in the neurosciences[@MullerPythonNeuroscience2015; @SchlaflyPythonpracticingneuroscientist2020], but the two main packages for spectral analysis in Python, `nitime` [@Rokem2020] and `mne-python` [@GramfortEtAl2013a], have issues that make them more difficult to use in many situations. For example, `nitime` implements several estimators of the power spectrum, but lacks spectrograms and windowed spectral estimators. `mne-python` is a much larger package designed as a full-featured analysis library for EEG and MEG data, and using the spectral analysis functions requires representing data using its ecosystem. Users of other data modalities, such as non-human electrophysiology data, may thus find much of `mne-python` inappropriate or unsuited for their analyses. This is an important problem because the non-human neurosciences are undergoing a period of great technological development; more and more signals are being collected simultaneously, and the duration of these signals is becoming longer as chronic recordings become possible. This rapid increase in the size and duration of datasets demands a lightweight, fast, and efficient spectral estimation package. `spectral_connectivity` is designed to handle multiple time series and can exploit GPUs for faster and more efficient computation. In addition, it can block compute important quantities such as the cross-spectral matrix in order to reduce memory burdens caused by large datasets. `spectral_connectivity` is also designed to be a lightweight package that has a simple user interface and can be easily be incorporated with other packages. Finally, `spectral_connectivity` also implements several connectivity measures that have not previously been implemented in Python such as the non-parametric version of the spectral granger causality and canonical coherence.
 
 `spectral_connectivity` has already shown its utility to the neuroscience field. The package has already been used in a number of publications and pre-prints in neuroscience [@KuhnertDetectionDirectedConnectivities2019; @VargaNetworkPathConvergence2021; @LauroSubthalamicCorticalNetwork2021;
[…]
