### 2022-10-14 Eden-Kramer-Lab/spectral_connectivity#32 @sappelhoff
@edeno I see you have started addressing the points I have raised: If you have finished with individual points, please feel free to close the respective issue that I opened and tag me, so I can double check.
