### 2022-10-14 fd647dd
Add path to lcov
