+++ .github/workflows/PR-test.yml
@@ -43,6 +43,7 @@ jobs:
       uses: coverallsapp/github-action@master
       with:
         github-token: ${{ secrets.GITHUB_TOKEN }}
+        path-to-lcov: coverage.lcov
     - name: Test notebooks
       shell: bash -l {0}
       run: |
