### 2022-10-14 Eden-Kramer-Lab/spectral_connectivity#32 @edeno
Okay thanks. @sappelhoff added .mailmap in 22281c7de5b06dd897eb5ddea15f8fc90f844ed5
