### 2022-10-14 Eden-Kramer-Lab/spectral_connectivity#43 @edeno
@sappelhoff I added more metadata. Please let me know if there are particular metadata that you feel are important to add.
