### 2022-10-14 Eden-Kramer-Lab/spectral_connectivity#39 @edeno
@sappelhoff I removed this CI step.
