### 2022-10-17 16d8f9c
Add code of conduct
Partially addresses issue #36
