### 2022-10-17 9a0e5ed
Add instructions for building documentation and making a release.
Address issue #36
