### 2022-10-17 Eden-Kramer-Lab/spectral_connectivity#36 @edeno
@sappelhoff 16d8f9c024de9ab1c4b219049306e90911322f36 and 9a0e5ed2d001f5a4c4f5dbf6de93aacd74e3f51f should address this issue.
