### 2022-10-17 ac925cf
Add authorship requirements
Address issue #36
