### 2022-10-17 Eden-Kramer-Lab/spectral_connectivity#41 @edeno
@sappelhoff addressed authorship in ac925cf44a3b9cb5d12b22ca22a8a20370b03557.
