### 2022-10-18 Eden-Kramer-Lab/spectral_connectivity#37 @sappelhoff
to be clear, this is not a requirement whatsoever, just a question / recommendation.
