+++ docs/conf.py
@@ -67,7 +67,7 @@ author = "Eric Denovellis"
 # The short X.Y version.
 version = "1..0"
 # The full version, including alpha/beta/rc tags.
-release = "1.0.0"
+release = "1.0.3"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
