### 2022-10-19 df6d928
Try to reformat docs and include notebooks
