+++ docs/conf.py
@@ -30,14 +30,17 @@ sys.path.insert(0, os.path.abspath("../"))
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    "sphinx.ext.autodoc",
+    "sphinx.ext.autodoc",  # Core Sphinx library for auto html doc generation from docstrings
+    "sphinx.ext.autosummary",  # Create neat summary tables for modules/classes/methods etc
     "sphinx.ext.coverage",
-    "sphinx.ext.intersphinx",
+    "sphinx.ext.intersphinx",  # Link to other project's documentation (see mapping below)
     "sphinx.ext.mathjax",
     "sphinx.ext.napoleon",
-    "sphinx.ext.viewcode",
+    "sphinx.ext.viewcode",  # Add a link to the Python source code for classes, functions etc.
+    "nbsphinx",  # Integrate Jupyter Notebooks and Sphinx
     "numpydoc",
 ]
+autosummary_generate = True
 
 add_module_names = False
 
