+++ docs/api.rst
@@ -1,49 +1,6 @@
-.. currentmodule:: spectral_connectivity
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-module-template.rst
+   :recursive:
 
-###
-API
-###
-
-spectral_connectivity package
-=============================
-
-spectral_connectivity.connectivity module
------------------------------------------
-
-.. automodule:: spectral_connectivity.connectivity
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-spectral_connectivity.minimum_phase_decomposition module
---------------------------------------------------------
-
-.. automodule:: spectral_connectivity.minimum_phase_decomposition
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-spectral_connectivity.statistics module
----------------------------------------
-
-.. automodule:: spectral_connectivity.statistics
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-spectral_connectivity.transforms module
----------------------------------------
-
-.. automodule:: spectral_connectivity.transforms
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-spectral_connectivity.wrapper module
----------------------------------------
-
-.. automodule:: spectral_connectivity.wrapper
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   spectral_connectivity
