+++ docs/tutorials.rst
@@ -1,0 +1,8 @@
+Tutorials
+=========
+
+.. toctree::
+   :caption: Introductory tutorial
+   :maxdepth: 1
+
+   .../examples/Tutorial_On_Simulated_Examples.ipynb
