+++ docs/requirements-docs.txt
@@ -2,3 +2,11 @@ spectral_connectivity
 numpydoc==1.1.0
 sphinx
 sphinx_rtd_theme==0.5.1
+pydata-sphinx-theme
+sphinx-autodoc-typehints
+nbsphinx
+ipython
+jupytext
+jupyter
+matplotlib
+numpy
