### 2022-10-19 1e7b337
Remove methods in docstring
