### 2022-10-19 1d06b03
Make asnumpy and non_negative_frequencies decorators private
