+++ spectral_connectivity/connectivity.py
@@ -42,8 +42,8 @@ EXPECTATION = {
 }
 
 
-def asnumpy(connectivity_measure):
-    """If cupy array, then return as numpy. Else return numpy array"""
+def _asnumpy(connectivity_measure):
+    """Decorator that transforms cupy array to numpy array. If cupy is not installed, then return original."""
 
     @wraps(connectivity_measure)
     def wrapper(*args, **kwargs):
@@ -59,7 +59,7 @@ def asnumpy(connectivity_measure):
     return wrapper
 
 
-def non_negative_frequencies(axis):
+def _non_negative_frequencies(axis):
     """Decorator that removes the negative frequencies."""
 
     def decorator(connectivity_measure):
@@ -157,14 +157,14 @@ class Connectivity:
         )
 
     @property
-    @asnumpy
-    @non_negative_frequencies(axis=0)
+    @_asnumpy
+    @_non_negative_frequencies(axis=0)
     def frequencies(self):
         if self._frequencies is not None:
             return self._frequencies
 
     @property
-    @asnumpy
+    @_asnumpy
     def all_frequencies(self):
         if self._frequencies is not None:
             return self._frequencies
@@ -246,7 +246,7 @@ class Connectivity:
         return minimum_phase_decomposition(self._expectation_cross_spectral_matrix())
 
     @property
-    @non_negative_frequencies(axis=-3)
+    @_non_negative_frequencies(axis=-3)
     def _transfer_function(self):
         return _estimate_transfer_function(self._minimum_phase_factor)
 
@@ -270,12 +270,12 @@ class Connectivity:
         else:
             return np.prod([self.fourier_coefficients.shape[axis] for axis in axes])
 
-    @asnumpy
-    @non_negative_frequencies(axis=-2)
+    @_asnumpy
+    @_non_negative_frequencies(axis=-2)
     def power(self):
         return self._power
 
-    @non_negative_frequencies(axis=-3)
+    @_non_negative_frequencies(axis=-3)
     def coherency(self):
         """The complex-valued linear association between time series in the
         frequency domain.
@@ -296,7 +296,7 @@ class Connectivity:
         complex_coherencey[..., diagonal_ind, diagonal_ind] = xp.nan
         return complex_coherencey
 
-    @asnumpy
+    @_asnumpy
     def coherence_phase(self):
         """The phase angle of the complex coherency.
 
@@ -307,7 +307,7 @@ class Connectivity:
         """
         return xp.angle(self.coherency())
 
-    @asnumpy
+    @_asnumpy
     def coherence_magnitude(self):
         """The magnitude of the complex coherency.
 
@@ -320,8 +320,8 @@ class Connectivity:
         """
         return _squared_magnitude(self.coherency())
 
-    @asnumpy
-    @non_negative_frequencies(axis=-3)
+    @_asnumpy
+    @_non_negative_frequencies(axis=-3)
     def imaginary_coherence(self):
         """The normalized imaginary component of the cross-spectrum.
 
@@ -490,8 +490,8 @@ class Connectivity:
         except AttributeError:
             return global_coherence, unnormalized_global_coherence
 
-    @asnumpy
-    @non_negative_frequencies(axis=-3)
+    @_asnumpy
+    @_non_negative_frequencies(axis=-3)
     def phase_locking_value(self):
         """The cross-spectrum with the power for each signal scaled to
         a magnitude of 1.
@@ -520,8 +520,8 @@ class Connectivity:
 
         return self._expectation_cross_spectral_matrix(fcn=fcn)
 
-    @asnumpy
-    @non_negative_frequencies(axis=-3)
+    @_asnumpy
+    @_non_negative_frequencies(axis=-3)
     def phase_lag_index(self):
         """A non-parametric synchrony measure designed to mitigate power
[…]
