+++ docs/requirements-docs.txt
@@ -1,4 +1,3 @@
-spectral_connectivity
 numpydoc==1.1.0
 sphinx
 sphinx_rtd_theme==0.5.1
