+++ docs/requirements-docs.txt
@@ -1,3 +1,4 @@
+spectral_connectivity
 numpydoc==1.1.0
 sphinx
 sphinx_rtd_theme==0.5.1
