### 2022-10-19 a0ccb0e
Make function private
