### 2022-10-19 2a0bdf8
Make dimensions lower case
