### 2022-10-19 Eden-Kramer-Lab/spectral_connectivity#38 @edeno
@sappelhoff function signatures are now fixed in the docs: https://spectral-connectivity.readthedocs.io/en/latest/_autosummary/spectral_connectivity.html
