+++ docs/tutorials.rst
@@ -5,4 +5,4 @@ Tutorials
    :caption: Introductory tutorial
    :maxdepth: 1
 
-   ../examples/Tutorial_On_Simulated_Examples.ipynb
+   examples/Tutorial_On_Simulated_Examples.ipynb
