### 2022-10-19 Eden-Kramer-Lab/spectral_connectivity#37 @edeno
I am happy to consider this move since it would make it easier to install. The reason for not doing it so far was mostly inertia. The package started on my own personal channel and it was easier to release things that way initially.
