### 2022-10-21 #4840 @sappelhoff
@emdupre I have completed my review. All issues that I raised have been addressed to my full satisfaction. I believe that `spectral_connectivity` is a valuable package that will serve the community well, as already evidenced in several publications. I recommend this paper for publication in JOSS.

For details, please see my checklist and all my linked issues and PRs in this review thread.
