### 2022-10-21 Eden-Kramer-Lab/spectral_connectivity#44 @EtienneCmb
[JOSS] Axes for expectation`time_trials_tapers`
Hi everyone,

For the review in JOSS  (https://github.com/openjournals/joss-reviews/issues/4840), could you check the axes used for averaging when expectation is `time_trials_tapers`, I think it should be (0, 1, 2) and not [(1, 2, 3)](https://github.com/Eden-Kramer-Lab/spectral_connectivity/blob/master/spectral_connectivity/connectivity.py#L41). Is it intentional to not mention this option in the doc of the Connectivity class?
