### 2022-10-21 5dc0b96
Fix expectation for time_trials_tapers
Issue #44
