### 2022-10-21 Eden-Kramer-Lab/spectral_connectivity#44 @edeno
Yes, that somehow snuck in. Thanks for catching that. Fixed by 5dc0b966b366c6e58bb7531486ef25d6b50d5131
