### 2022-10-21 Eden-Kramer-Lab/spectral_connectivity#45 @EtienneCmb
[JOSS] PLV returns complex values
The `phase_locking_value` method returns complex value because the `np.abs` is missing. I guess you made this choice because it's then used in the ppc but from the user side, especially for non-expert, it might be confusing.

A quick fix could be to have a `_phase_locking_value` without the absolute value and with :
```python
def phase_locking_value()
    return xp.abs(self._phase_locking_value())
```

And the ppc could use `_phase_locking_value` method instead. What do you think?

JOSS reference (https://github.com/openjournals/joss-reviews/issues/4840),
