### 2022-10-24 Eden-Kramer-Lab/spectral_connectivity#46 @EtienneCmb
[JOSS] Comparison with MNE
Regarding the comparison with MNE-Python, you wrote :

> mne-python is a much larger package designed as a full-featured analysis library for EEG and MEG data, and using the spectral analysis functions requires representing data using its ecosystem.

I understand your point, it's true that MNE strongly relies on its `Epoch` or `Raw` container that have not been designed to handle non-human electrophysiological data. But to be fair on the comparison, MNE also have functions that can be applied to non-human LFP using standard NumPy array inputs. For example, in line with this package :
- [mne.time_frequency.psd_array_multitaper](https://mne.tools/stable/generated/mne.time_frequency.psd_array_multitaper.html) and [mne.time_frequency.tfr_array_multitaper](https://mne.tools/stable/generated/mne.time_frequency.tfr_array_multitaper.html)  : PSD and time-frequency maps using multi-tapers
- [mne_connectivity.spectral_connectivity_time](https://mne.tools/mne-connectivity/stable/generated/mne_connectivity.spectral_connectivity_time.html#mne_connectivity.spectral_connectivity_time) and [mne_connectivity.spectral_connectivity_epochs](https://mne.tools/mne-connectivity/stable/generated/mne_connectivity.spectral_connectivity_epochs.html#mne_connectivity.spectral_connectivity_epochs) : both functions implement connectivity metrics, using multi-tapers and support on NumPy array

In my opinion, the strength of this package lies in the parameter `expectation_type`. It's a smart and flexible design because it allows having single-trial and possibly dynamic estimations of functional connectivity using a single implementation, while in MNE they had to use two different functions.

xref : JOSS (https://github.com/openjournals/joss-reviews/issues/4840)
