### 2022-10-26 Eden-Kramer-Lab/spectral_connectivity#45 @edeno
That seems like a sensible change. The PLV can be used directionally so think I was aiming to leave it flexible, but people probably expect it to be the undirected version.
