### 2022-10-26 649db3e
PLV returns magnitude only
Addresses issue #45
