### 2022-10-27 Eden-Kramer-Lab/spectral_connectivity#45 @edeno
@EtienneCmb should be fixed in 649db3e2a595862f771ac5b4585f33d3dfff8c33
