### 2022-11-02 Eden-Kramer-Lab/spectral_connectivity#46 @emilyps14
True, although it looks like `mne_connectivity.spectral_connectivity_time` does require an `Epochs` object, as it accesses `data.ch_names` etc.

How about I soften the language in that sentence, to say something like:
> `mne-python` is a much larger package designed as a full-featured analysis library for EEG and MEG data, and works best when the data is represented using its ecosystem (i.e. `Epochs` and `Raw` objects). While some of the spectral connectivity functionality can work with `array_like` objects, users of other data modalities such as non-human electrophysiology data may find `mne-python` too cumbersome for their application.

Then we can highlight the `expectation_type` functionality in a footnote:
> ...`spectral_connectivity` is designed to handle multiple time series flexibly[^1] and can exploit GPUs for faster and more efficient computation. ...
[^1]: for example, the `expectation_type` parameter of the `Connectivity` class allows the user to average over trials, tapers, or both, allowing the same module to yield single-trial or summarized estimates.
