### 2022-11-02 e8083c1
Editing paper comparison to MNE-python
