### 2022-11-02 Eden-Kramer-Lab/spectral_connectivity#46 @EtienneCmb
Hi @emilyps14,

> True, although it looks like mne_connectivity.spectral_connectivity_time does require an Epochs object, as it accesses data.ch_names etc.

You're right, this function is still WIP. When this [PR](https://github.com/mne-tools/mne-connectivity/pull/104) will be merged, `array_like` inputs are going to be supported as well (see [here](https://github.com/mne-tools/mne-connectivity/blob/7c64633b5c8bd1973b98214c4e3d253fd6ed5095/mne_connectivity/spectral/time.py#L214), it's already implemented). The PR is under review, it's a matter of a few weeks before being merged.

> "[...] users of other data modalities such as non-human electrophysiology data may find mne-python too cumbersome for their application."

`array_like` inputs can be converted to `Epoch` in a very few lines of code :

```python
from mne import EpochsArray, create_info

info = create_info(ch_names, sfreq)
epoch = EpochsArray(data, info, tmin=times[0])
```

I agree with you that it's not as direct as your package but I would not qualify it as "cumbersome".
