### 2022-11-02 Eden-Kramer-Lab/spectral_connectivity#46 @emilyps14
It's cumbersome also because of the size of the package, and having to install the whole package (which is nontrivial) just to use a small set of functions which don't leverage the primary strengths of MNE-python. E.g. the installers recommend having a conda environment just for MNE-python, which doesn't make sense for folks who only want to use it for spectral connectivity functions.
