### 2022-11-02 Eden-Kramer-Lab/spectral_connectivity#46 @EtienneCmb
Personal experience, but all the students here are using MNE, none ever had any issue installing MNE on a broad range of systems. Regarding the conda environment, it's convenient for non-programmers as everything (data analysis, visualisation etc.) can be installed via a single command and avoid any conflicts between versions. But if folks are just interested in spectral connectivity functions, MNE can be installed through [pip](https://github.com/mne-tools/mne-python#installing-mne-python) and the list of [minimum requirements](https://github.com/mne-tools/mne-python#dependencies) is light.

IMO, you could consider different points of comparison with MNE :
1) The architecture of your code is more flexible (cf. `expectation_type`)
2) You included connectivity metrics they don't have (e.g. the coherence that can combine several channels. This feature is indeed interesting for sparse recordings like ECoG, iEEG, LFP etc.)
3) Maybe your implementation is faster or more accurate, in that case, you could provide a comparison
