### 2022-11-07 #4840 @emdupre
👋 Hi @AJQuinn, I wanted to make sure you weren't facing issues in getting started on your review ?
Please let me know if there's anything I can do to help, and thank you again for agreeing to review Spectral Connectivity !
