### 2022-11-19 #4840 @emdupre
👋 Hi @EtienneCmb, I just wanted to check in and see how this review is going for you. Please let me know if I can answer any questions that might help in finalizing your review ! And thanks again for agreeing to review Spectral connectivity for _JOSS_.
