### 2022-11-27 Eden-Kramer-Lab/spectral_connectivity#50 @AJQuinn
It would also be great to include more information on the docs page - some of this is duplicated in the README file but should be consistent across github and readthedocs if both are maintained. Currently a user landing on readthedocs could get a bit lost...

Can you look into adding/fixing:

- [ ] Details about bug reporting or making code contributions
- [ ] Module level docstrings to add more info on the API pages: https://spectral-connectivity.readthedocs.io/en/latest/_autosummary/spectral_connectivity.html
- [ ] The jupyter tutorial pages seem to be empy - should be linked to relevant pages: https://spectral-connectivity.readthedocs.io/en/latest/tutorials.html
- [ ] Grammar and capitalisation on landing page aren't great, seems like points should be rendered as a list perhaps? https://spectral-connectivity.readthedocs.io/en/latest/index.html
