### 2022-11-27 Eden-Kramer-Lab/spectral_connectivity#51 @AJQuinn
[JOSS] Usage examples and tutorials
Hello - I like the usage example in the main README but the tutorial jupyter notebooks are most likely only helpful for current experts. The content is good but contains very large code blocks and reads a bit like a list.

I would strongly recommend including one or two tutorials which break down the large code blocks and include a good amount of prose explaning what each chunk of code is doing and why. This doesn't need to be absolutely extensive but is hugely valuable for new users.

Some examples from similar toolboxes

- https://sails.readthedocs.io/en/stable/tutorials/tutorial5.html
- https://mne.tools/mne-connectivity/stable/auto_examples/dpli_wpli_pli.html
- https://nipy.org/nitime/examples/ar_est_2vars.html

xref: https://github.com/openjournals/joss-reviews/issues/4840
