### 2022-11-27 #4840 @AJQuinn
Hi @emdupre @edeno - Apologies for the long delay, my review is in progress and should be completed tomorrow. I'm impressed overall. This is a useful package though I think the documentation needs a bit of work.

I've added some issues on the docs and will complete some comments on the code tomorrow.
