### 2022-11-28 Eden-Kramer-Lab/spectral_connectivity#51 @edeno
Got it. Agreed that there needs to be more intro to using the API and I think the simplicity of using the API is one of our strengths. I will work on the documentation website and provide more introductory tutorials.
