### 2022-11-30 Eden-Kramer-Lab/spectral_connectivity#51 @AJQuinn
Great - agreed that the API is a strength, this really comes across in the README example.
