### 2022-11-30 Eden-Kramer-Lab/spectral_connectivity#52 @AJQuinn
Perhaps thats too much for a single comment..! Please let me know if that all makes sense, happy to chat more if anything is worth clarifying or discussing.
