### 2022-11-30 99ec219
Remove Makefile
Removed Makefile per #52 . Was not being used.
