### 2022-11-30 2fdd324
Add docstrings
per issue #52
