### 2022-11-30 83c37c6
Add black code linting
Formalize codestyle. Per issue #52 .
