### 2022-11-30 6943b79
Remove stub
