+++ tests/test_transforms.py
@@ -2,13 +2,18 @@ import numpy as np
 from nitime.algorithms.spectral import dpss_windows as nitime_dpss_windows
 from pytest import mark
 from scipy.signal import correlate
-from spectral_connectivity.transforms import (Multitaper, _add_axes,
-                                              _auto_correlation,
-                                              _fix_taper_sign,
-                                              _get_low_bias_tapers,
-                                              _get_taper_eigenvalues,
-                                              _multitaper_fft, _sliding_window,
-                                              dpss_windows)
+
+from spectral_connectivity.transforms import (
+    Multitaper,
+    _add_axes,
+    _auto_correlation,
+    _fix_taper_sign,
+    _get_low_bias_tapers,
+    _get_taper_eigenvalues,
+    _multitaper_fft,
+    _sliding_window,
+    dpss_windows,
+)
 
 
 def test__add_axes():
@@ -31,96 +36,107 @@ def test__add_axes():
 
 
 @mark.parametrize(
-    'test_array, window_size, step_size, axis, expected_array',
-    [(np.arange(1, 6), 3, 1, -1, np.array([[1, 2, 3],
-                                           [2, 3, 4],
-                                           [3, 4, 5]])),
-     (np.arange(1, 6), 3, 2, -1, np.array([[1, 2, 3],
-                                           [3, 4, 5]])),
-     (np.arange(0, 6).reshape((2, 3)), 2, 1, 0, np.array([[[0, 3],
-                                                           [1, 4],
-                                                           [2, 5]]]))
-     ])
-def test__sliding_window(
-        test_array, window_size, step_size, axis, expected_array):
+    "test_array, window_size, step_size, axis, expected_array",
+    [
+        (np.arange(1, 6), 3, 1, -1, np.array([[1, 2, 3], [2, 3, 4], [3, 4, 5]])),
+        (np.arange(1, 6), 3, 2, -1, np.array([[1, 2, 3], [3, 4, 5]])),
+        (
+            np.arange(0, 6).reshape((2, 3)),
+            2,
+            1,
+            0,
+            np.array([[[0, 3], [1, 4], [2, 5]]]),
+        ),
+    ],
+)
+def test__sliding_window(test_array, window_size, step_size, axis, expected_array):
     assert np.allclose(
         _sliding_window(
-            test_array, window_size=window_size, step_size=step_size,
-            axis=axis),
-        expected_array)
+            test_array, window_size=window_size, step_size=step_size, axis=axis
+        ),
+        expected_array,
+    )
 
 
 @mark.parametrize(
-    'time_halfbandwidth_product, expected_n_tapers',
-    [(3, 5), (1, 1), (1.75, 2)])
+    "time_halfbandwidth_product, expected_n_tapers", [(3, 5), (1, 1), (1.75, 2)]
+)
 def test_n_tapers(time_halfbandwidth_product, expected_n_tapers):
     n_time_samples, n_trials, n_signals = 100, 10, 2
     time_series = np.zeros((n_time_samples, n_trials, n_signals))
     m = Multitaper(
-        time_series=time_series,
-        time_halfbandwidth_product=time_halfbandwidth_product)
+        time_series=time_series, time_halfbandwidth_product=time_halfbandwidth_product
+    )
     assert m.n_tapers == expected_n_tapers
 
 
 @mark.parametrize(
-    'sampling_frequency, time_window_duration, expected_duration',
[…]
