+++ tests/test_statistics.py
@@ -1,10 +1,13 @@
 import numpy as np
 from pytest import mark
-from spectral_connectivity.statistics import (Benjamini_Hochberg_procedure,
-                                              Bonferroni_correction,
-                                              coherence_bias,
-                                              fisher_z_transform,
-                                              get_normal_distribution_p_values)
+
+from spectral_connectivity.statistics import (
+    Benjamini_Hochberg_procedure,
+    Bonferroni_correction,
+    coherence_bias,
+    fisher_z_transform,
+    get_normal_distribution_p_values,
+)
 
 
 def test_get_normal_distribution_p_values():
@@ -18,39 +21,39 @@ def test_fisher_z_transform():
     bias1, bias2 = 3, 6
     expected_difference_z = np.ones((2, 2))
     assert np.allclose(
-        fisher_z_transform(
-            coherency, bias1, coherency2=coherency, bias2=bias2),
-        expected_difference_z)
+        fisher_z_transform(coherency, bias1, coherency2=coherency, bias2=bias2),
+        expected_difference_z,
+    )
 
 
 @mark.parametrize(
-    'p_values, expected_is_significant',
-    [(np.ones((10, 2)), np.zeros((10, 2), dtype=bool)),
-     (np.zeros((10, 2)), np.ones((10, 2), dtype=bool)),
-     (np.array([0.03, 0.01, 0.04, 0.06]),
-      np.array([False, True, False, False])),
-     (np.array([0.03, 0.01, 0.04, 0.05]),
-      np.ones((4,)))])
+    "p_values, expected_is_significant",
+    [
+        (np.ones((10, 2)), np.zeros((10, 2), dtype=bool)),
+        (np.zeros((10, 2)), np.ones((10, 2), dtype=bool)),
+        (np.array([0.03, 0.01, 0.04, 0.06]), np.array([False, True, False, False])),
+        (np.array([0.03, 0.01, 0.04, 0.05]), np.ones((4,))),
+    ],
+)
 def test_Benjamini_Hochberg_procedure(p_values, expected_is_significant):
     alpha = 0.05
     assert np.allclose(
-        Benjamini_Hochberg_procedure(p_values, alpha),
-        expected_is_significant)
+        Benjamini_Hochberg_procedure(p_values, alpha), expected_is_significant
+    )
 
 
 @mark.parametrize(
-    'p_values, expected_is_significant',
-    [(np.ones((10, 2)), np.zeros((10, 2), dtype=bool)),
-     (np.zeros((10, 2)), np.ones((10, 2), dtype=bool)),
-     (np.array([0.03, 0.01, 0.04, 0.06]),
-      np.array([False, True, False, False])),
-     (np.array([0.03, 0.01, 0.04, 0.05]),
-      np.array([False, True, False, False]))])
+    "p_values, expected_is_significant",
+    [
+        (np.ones((10, 2)), np.zeros((10, 2), dtype=bool)),
+        (np.zeros((10, 2)), np.ones((10, 2), dtype=bool)),
+        (np.array([0.03, 0.01, 0.04, 0.06]), np.array([False, True, False, False])),
+        (np.array([0.03, 0.01, 0.04, 0.05]), np.array([False, True, False, False])),
+    ],
+)
 def test_Bonferroni_correction(p_values, expected_is_significant):
     alpha = 0.05
-    assert np.allclose(
-        Bonferroni_correction(p_values, alpha),
-        expected_is_significant)
+    assert np.allclose(Bonferroni_correction(p_values, alpha), expected_is_significant)
 
 
 def test_coherence_bias():
