### 2022-11-30 e5dc69b
Match setup.py dependency requirements
Per issue #52
