+++ environment.yml
@@ -3,7 +3,8 @@ channels:
 - conda-forge
 - defaults
 dependencies:
-- numpy
+- numpy >= 1.11
+- pandas >= 0.18.0
 - matplotlib
 - scipy
 - xarray
