### 2022-11-30 d1a1985
Update docstring
Per issue #52
