+++ docs/index.rst
@@ -22,7 +22,7 @@ Tutorials
 =========
 .. toctree::
    :maxdepth: 1
-   ../examples/Tutorial_On_Simulated_Examples.ipynb
+   ../../examples/Tutorial_On_Simulated_Examples.ipynb
 
 
 API documentation
