### 2022-12-01 fb2a3e2
Try pydata sphinx theme
