+++ docs/conf.py
@@ -39,6 +39,7 @@ extensions = [
     "sphinx.ext.viewcode",  # Add a link to the Python source code for classes, functions etc.
     "nbsphinx",  # Integrate Jupyter Notebooks and Sphinx
     "numpydoc",
+    "myst_nb",
 ]
 autosummary_generate = True
 
@@ -122,7 +123,7 @@ todo_include_todos = False
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = "sphinx_rtd_theme"
+html_theme = "pydata_sphinx_theme"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
