+++ docs/requirements-docs.txt
@@ -3,6 +3,7 @@ numpydoc==1.1.0
 sphinx
 sphinx_rtd_theme==0.5.1
 pydata-sphinx-theme
+myst_nb
 sphinx-autodoc-typehints
 nbsphinx
 ipython
