### 2022-12-01 d90ccfa
Remove readthedocs theme
