+++ docs/requirements-docs.txt
@@ -1,7 +1,6 @@
 spectral_connectivity
 numpydoc==1.1.0
 sphinx
-sphinx_rtd_theme==0.5.1
 pydata-sphinx-theme
 myst_nb
 sphinx-autodoc-typehints
