+++ docs/index.md
@@ -1,0 +1,153 @@
+# spectral_connectivity
+
+[![PR Test](https://github.com/Eden-Kramer-Lab/spectral_connectivity/actions/workflows/PR-test.yml/badge.svg)](https://github.com/Eden-Kramer-Lab/spectral_connectivity/actions/workflows/PR-test.yml)
+[![DOI](https://zenodo.org/badge/104382538.svg)](https://zenodo.org/badge/latestdoi/104382538)
+[![Binder](https://mybinder.org/badge.svg)](https://mybinder.org/v2/gh/Eden-Kramer-Lab/spectral_connectivity/master)
+[![status](https://joss.theoj.org/papers/27eb33e699c9ea723783c44576d081bb/status.svg)](https://joss.theoj.org/papers/27eb33e699c9ea723783c44576d081bb)
+[![PyPI version](https://badge.fury.io/py/spectral_connectivity.svg)](https://badge.fury.io/py/spectral_connectivity)
+[![Anaconda-Server Badge](https://anaconda.org/edeno/spectral_connectivity/badges/version.svg)](https://anaconda.org/edeno/spectral_connectivity)
+[![Documentation Status](https://readthedocs.org/projects/spectral-connectivity/badge/?version=latest)](https://spectral-connectivity.readthedocs.io/en/latest/?badge=latest)
+[![Coverage Status](https://coveralls.io/repos/github/Eden-Kramer-Lab/spectral_connectivity/badge.svg?branch=master)](https://coveralls.io/github/Eden-Kramer-Lab/spectral_connectivity?branch=master)
+
+[**Tutorials**](#tutorials)
+| [**Documentation**](#documentation)
+| [**Usage Example**](#usage-example)
+| [**Installation**](#installation)
+| [**Developer Installation**](#developer-installation)
+
+### What is spectral_connectivity?
+
+`spectral_connectivity` is a Python software package that computes multitaper spectral estimates and frequency-domain brain connectivity measures such as coherence, spectral granger causality, and the phase lag index using the multitaper Fourier transform. Although there are other Python packages that do this (see [nitime](https://github.com/nipy/nitime) and [MNE-Python](https://github.com/mne-tools/mne-python)), `spectral_connectivity` has several differences:
+
++ it is designed to handle multiple time series at once
++ it caches frequently computed quantities such as the cross-spectral matrix and minimum-phase-decomposition, so that connectivity measures that use the same processing steps can be more quickly computed.
++ it decouples the time-frequency transform and the connectivity measures so that if you already have a preferred way of computing Fourier coefficients (i.e. from a wavelet transform), you can use that instead.
++ it implements the non-parametric version of the spectral granger causality in Python.
++ it implements the canonical coherence, which can
+efficiently summarize brain-area level coherences from multielectrode recordings.
++ easier user interface for the multitaper fourier transform
++ all function are GPU-enabled if `cupy` is installed and the environmental variable `SPECTRAL_CONNECTIVITY_ENABLE_GPU` is set to 'true'.
+
+### Tutorials
+
+See the following notebooks for more information on how to use the package:
+
++ [Tutorial on simulated examples](examples/Tutorial_On_Simulated_Examples.ipynb)
++ [Tutorial using paper examples](examples/Tutorial_Using_Paper_Examples.ipynb)
+
+### Usage Example
+
+```python
+from spectral_connectivity import Multitaper, Connectivity
+
+# Compute multitaper spectral estimate
+m = Multitaper(time_series=signals,
+               sampling_frequency=sampling_frequency,
+               time_halfbandwidth_product=time_halfbandwidth_product,
[…]
