### 2022-12-01 e11796c
Remove index.rst
