+++ docs/requirements-docs.txt
@@ -10,5 +10,3 @@ jupytext
 jupyter
 matplotlib
 numpy
-myst_parser
-myst_nb
