+++ docs/index.md
@@ -148,6 +148,4 @@ that is rescued by antipsychotic drugs <https://doi.org/10.1101/2021.02.03.42958
 :maxdepth: 2
 :hidden:
 _autosummary/spectral_connectivity
-genindex
-search
 ```
