+++ docs/requirements-docs.txt
@@ -1,8 +1,14 @@
 spectral_connectivity
 numpydoc==1.1.0
 sphinx
-pydata-sphinx-theme
+docutils
 myst_nb
+MyST-Parser
+pydata-sphinx-theme
+sphinx-autobuild
+sphinx-copybutton
+sphinx-design
+sphinx-examples
 sphinx-autodoc-typehints
 nbsphinx
 ipython
