### 2022-12-01 #4840 @AJQuinn
Hi @emdupre my review is finished. I'm impressed with the package. It works well and the code is high quality. There are some similar packages with overlapping functionality but I think `spectral_connectivity` is a strong addition to the field with several unique features. I felt parts of the documentation were lacking a bit and have outlined some suggestions for this in the linked comments. Happy to proceed once these are actioned.
