+++ docs/api.rst
@@ -1,3 +1,6 @@
+API reference
+=============
+
 .. autosummary::
    :toctree: _autosummary
    :template: custom-module-template.rst
