+++ docs/index.md
@@ -137,5 +137,5 @@ that is rescued by antipsychotic drugs <https://doi.org/10.1101/2021.02.03.42958
 ```{toctree}
 :maxdepth: 2
 :hidden:
-_autosummary/spectral_connectivity
+api
 ```
