+++ docs/api.rst
@@ -1,6 +1,3 @@
-API
-===
-
 .. autosummary::
    :toctree: _autosummary
    :template: custom-module-template.rst
