+++ docs/index.md
@@ -137,5 +137,6 @@ that is rescued by antipsychotic drugs <https://doi.org/10.1101/2021.02.03.42958
 ```{toctree}
 :maxdepth: 2
 :hidden:
+../examples/Tutorial_On_Simulated_Examples.ipynb
 api
 ```
