### 2022-12-01 cef9edb
Revert "Try different order"
This reverts commit 6a516c772e50cf6ed20d855e27c65974385ff066.
