+++ docs/api.rst
@@ -3,8 +3,4 @@
    :template: custom-module-template.rst
    :recursive:
 
-   spectral_connectivity.transforms
-   spectral_connectivity.connectivity
-   spectral_connectivity.statistics
-   spectral_connectivity.wrapper
-   spectral_connectivity.minimum_phase_decomposition
+   spectral_connectivity
