### 2022-12-01 c05310d
Use ipykernel instead of nb_conda
