+++ docs/requirements-docs.txt
@@ -18,3 +18,4 @@ jupyterlab
 matplotlib
 numpy
 ipykernel
+nb_conda
