+++ docs/conf.py
@@ -62,8 +62,8 @@ master_doc = "index"
 
 # General information about the project.
 project = "spectral_connectivity"
-copyright = "2022, Eric Denovellis"
-author = "Eric Denovellis"
+copyright = "2022, Eric L. Denovellis"
+author = "Eric L. Denovellis"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -79,7 +79,7 @@ release = "1.0.3"
 #
 # This is also used if you do content translation via gettext catalogs.
 # Usually you set "language" from the command line for these cases.
-language = None
+language = "En"
 
 # There are two options for replacing |today|: either, you set today to some
 # non-false value, then it is used:
