### 2022-12-01 848ac25
Revert "Add nbconda"
This reverts commit c70bda0f3f9eadf71197c3fa76d28e4139bd81d6.
