+++ docs/requirements-docs.txt
@@ -10,11 +10,9 @@ sphinx-copybutton
 sphinx-design
 sphinx-examples
 sphinx-autodoc-typehints
-nbsphinx
 ipython
 jupytext
 jupyter
-jupyterlab
 matplotlib
 numpy
 ipykernel
