### 2022-12-01 13611f4
Revert "remove nbsphinx"
This reverts commit 87dc43bc7c666c0247b0c65c7653a1ac4925753c.
