+++ docs/index.md
@@ -136,11 +136,6 @@ that is rescued by antipsychotic drugs <https://doi.org/10.1101/2021.02.03.42958
 
 ```{toctree}
 ---
-:caption: Contents:
-:maxdepth: 2
----
-examples/Tutorial_On_Simulated_Examples
----
 :caption: API:
 :maxdepth: 2
 ---
