+++ docs/conf.py
@@ -51,7 +51,7 @@ templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
-source_suffix = [".rst", ".md"]
+source_suffix = {".rst": "restructuredtext", ".myst": "myst-nb", ".ipynb": "myst-nb"}
 # source_suffix = ".rst"
 
 # The encoding of source files.
@@ -323,9 +323,10 @@ intersphinx_mapping = {
 
 # MyST
 myst_enable_extensions = [
-    "colon_fence",
+    "dollarmath",
+    "amsmath",
     "deflist",
-    "html_image",
+    "colon_fence",
 ]
 
 # MyST-NB
