+++ docs/api.rst
@@ -1,5 +1,9 @@
+API
+===
+
 .. autosummary::
    :toctree: _autosummary
+   :caption: API
    :template: custom-module-template.rst
    :recursive:
 
