### 2022-12-01 4242c07
Remove unused text
