### 2022-12-01 d300366
Revert "Temp remove notebooks"
This reverts commit bd8caf3f4aa8915d07a11561079122124ae774fe.
