### 2022-12-01 8078db7
Remove caption
