+++ docs/api.rst
@@ -3,7 +3,6 @@ API
 
 .. autosummary::
    :toctree: _autosummary
-   :caption: API
    :template: custom-module-template.rst
    :recursive:
 
