+++ docs/conf.py
@@ -10,6 +10,14 @@ import sys
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 sys.path.insert(0, os.path.abspath("../"))
 
+# -- Project information -----------------------------------------------------
+
+master_doc = "index"
+
+project = "spectral_connectivity"
+copyright = "2022, Eric L. Denovellis"
+author = "Eric L. Denovellis"
+
 # -- General configuration ------------------------------------------------
 
 extensions = [
@@ -39,13 +47,6 @@ source_suffix = {".rst": "restructuredtext", ".myst": "myst-nb", ".ipynb": "myst
 # The encoding of source files.
 # source_encoding = 'utf-8-sig'
 
-# The master toctree document.
-master_doc = "index"
-
-# General information about the project.
-project = "spectral_connectivity"
-copyright = "2022, Eric L. Denovellis"
-author = "Eric L. Denovellis"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -66,7 +67,7 @@ language = "En"
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This patterns also effect to html_static_path and html_extra_path
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "**.ipynb_checkpoints"]
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = "sphinx"
