### 2022-12-01 2a8f047
Revert "Revert "Temp remove notebooks""
This reverts commit d30036665d3047bb3754769be4782c44e3dd3104.
