### 2022-12-01 611ed1b
Revert "Only do the modules"
This reverts commit cef4710e08ea84051e1ee0a9ff092b3cbaed3ad4.
