+++ docs/api.rst
@@ -1,14 +1,9 @@
 API
 ===
-Here you can find all the classes and methods included in `spectral_connectivity`.
 
 .. autosummary::
    :toctree: _autosummary
    :template: custom-module-template.rst
    :recursive:
 
-   spectral_connectivity/connectivity
-   spectral_connectivity/minimum_phase_decomposition
-   spectral_connectivity/statistics
-   spectral_connectivity/transforms
-   spectral_connectivity/wrapper
+   spectral_connectivity
