+++ docs/index.md
@@ -153,6 +153,6 @@ that is rescued by antipsychotic drugs <https://doi.org/10.1101/2021.02.03.42958
 :titlesonly:
 :maxdepth: 1
 
-api
+_autosummary/spectral_connectivity
 contributing
 ```
