+++ docs/api.rst
@@ -1,9 +1,15 @@
 API
 ===
 
+.. automodule:: spectral_connectivity
+
 .. autosummary::
-   :toctree: _autosummary
+   :toctree:
    :template: custom-module-template.rst
    :recursive:
 
-   spectral_connectivity
+   spectral_connectivity.connectivity
+   spectral_connectivity.minimum_phase_decomposition
+   spectral_connectivity.statistics
+   spectral_connectivity.transforms
+   spectral_connectivity.wrapper
