+++ docs/index.md
@@ -145,6 +145,8 @@ that is rescued by antipsychotic drugs <https://doi.org/10.1101/2021.02.03.42958
 :hidden:
 :maxdepth: 2
 
+examples/Tutorial_On_Simulated_Examples
+
 ```
 
 ```{toctree}
