+++ docs/index.md
@@ -146,6 +146,7 @@ that is rescued by antipsychotic drugs <https://doi.org/10.1101/2021.02.03.42958
 :maxdepth: 2
 
 examples/Tutorial_On_Simulated_Examples
+examples/Tutorial_Using_Paper_Examples
 
 ```
 
