+++ docs/conf.py
@@ -90,6 +90,8 @@ html_static_path = ["_static"]
 # Output file base name for HTML help builder.
 htmlhelp_basename = "spectral_connectivitydoc"
 
+html_sidebars = {"**": ["sidebar-nav-bs"]}
+
 # -- Options for LaTeX output ---------------------------------------------
 
 latex_elements = {
