+++ docs/tutorials.rst
@@ -2,7 +2,8 @@ Tutorials
 =========
 
 .. toctree::
-   :caption: Introductory tutorial
    :maxdepth: 1
+   :includehidden:
 
-   examples/Tutorial_On_Simulated_Examples.ipynb
+   examples/Tutorial_On_Simulated_Examples
+   examples/Tutorial_Using_Paper_Examples
