### 2022-12-02 c368f54
Remove word documentation
