### 2022-12-02 da74008
Fix shape in docstrings
