+++ spectral_connectivity/connectivity.py
@@ -98,9 +98,7 @@ class Connectivity:
 
     Attributes
     ----------
-    fourier_coefficients : array, shape (n_time_windows, n_trials,
-                                         n_tapers, n_fft_samples,
-                                         n_signals)
+    fourier_coefficients : array, shape (n_time_windows, n_trials, n_tapers, n_fft_samples, n_signals)
         The compex-valued coefficients from a fourier transform. Note that
         this is expected to be the two-sided fourier coefficients
         (both the positive and negative lags). This is needed for the
@@ -140,9 +138,7 @@ class Connectivity:
 
         Parameters
         ----------
-        fourier_coefficients : np.ndarray, shape (n_time_windows, n_trials,
-                                         n_tapers, n_fft_samples,
-                                         n_signals)
+        fourier_coefficients : np.ndarray, shape (n_time_windows, n_trials, n_tapers, n_fft_samples, n_signals)
             The compex-valued coefficients from a fourier transform. Note that
             this is expected to be the two-sided fourier coefficients
             (both the positive and negative lags). This is needed for the
@@ -218,9 +214,7 @@ class Connectivity:
 
         Returns
         -------
-        cross_spectral_matrix : array, shape (n_time_windows, n_trials,
-                                              n_tapers, n_fft_samples,
-                                              n_signals, n_signals)
+        cross_spectral_matrix : array, shape (n_time_windows, n_trials, n_tapers, n_fft_samples, n_signals, n_signals)
 
         """
         fourier_coefficients = self.fourier_coefficients[..., xp.newaxis]
@@ -320,8 +314,7 @@ class Connectivity:
 
         Returns
         -------
-        complex_coherency : array, shape (..., n_fft_samples, n_signals,
-                                          n_signals)
+        complex_coherency : array, shape (..., n_fft_samples, n_signals, n_signals)
 
         """
         norm = xp.sqrt(
@@ -373,8 +366,7 @@ class Connectivity:
 
         Returns
         -------
-        imaginary_coherence_magnitude : array, shape (..., n_fft_samples,
-                                                      n_signals, n_signals)
+        imaginary_coherence_magnitude : array, shape (..., n_fft_samples, n_signals, n_signals)
 
         References
         ----------
@@ -403,8 +395,7 @@ class Connectivity:
 
         Returns
         -------
-        canonical_coherence : array, shape (n_time_samples, n_fft_samples,
-                                            n_groups, n_groups)
+        canonical_coherence : array, shape (n_time_samples, n_fft_samples, n_groups, n_groups)
             The maximimal coherence for each group pair
         labels : array, shape (n_groups,)
             The sorted unique group labels that correspond to `n_groups`
@@ -477,10 +468,7 @@ class Connectivity:
                                            n_fft_samples,
                                            n_components)
             The vector of global coherences (square of the singular values)
-        unnormalized_global_coherence : ndarray, shape (n_time_windows,
-                                                        n_fft_samples,
-                                                        n_signals,
-                                                        n_components)
+        unnormalized_global_coherence : ndarray, shape (n_time_windows, n_fft_samples, n_signals, n_components)
             The (unnormalized) global coherence vectors
 
         References
@@ -542,8 +530,7 @@ class Connectivity:
 
         Returns
         -------
-        phase_locking_value : array, shape (..., n_fft_samples, n_signals,
-                                            n_signals)
+        phase_locking_value : array, shape (..., n_fft_samples, n_signals, n_signals)
 
         References
         ----------
@@ -580,8 +567,7 @@ class Connectivity:
 
         Returns
         -------
[…]
