+++ docs/api.rst
@@ -1,5 +1,5 @@
-API
-===
+API Reference
+=============
 
 .. automodule:: spectral_connectivity
 
