+++ docs/tutorials.rst
@@ -1,5 +1,6 @@
 Tutorials
 =========
+
 .. toctree::
    :caption: Introductory tutorial
    :maxdepth: 1
