### 2022-12-02 Eden-Kramer-Lab/spectral_connectivity#53 @EtienneCmb
[JOSS] GPU computations
I tested GPU computations but I'm not sure I did it correctly as I couldn't observe differences in computing time (see [notebook here](https://colab.research.google.com/drive/1dA0GARhhMQgZ3ey56RnNh3F-huLfG-B6?usp=sharing)). Do you have an idea why?

Also, in the doc and in the `Multitaper` file, you define an environmental variable `SPECTRAL_CONNECTIVITY_ENABLE_GPU` but in the `Connectivity` file it's `SPEC_CON_ENABLE_GPU`.

xref : JOSS (https://github.com/openjournals/joss-reviews/issues/4840)
