### 2022-12-02 Eden-Kramer-Lab/spectral_connectivity#46 @EtienneCmb
Hi @emilyps14,

Regarding the comparison with other software, in our team, we also made a software called Frites (see [JOSS](https://doi.org/10.21105/joss.03842) paper) that includes single-trial estimates of FC, however, we are more on the information-theory side (see here the [API](https://brainets.github.io/frites/api/api_connectivity.html#connectivity-metrics)). Also, it's not as flexible as in your design.
