### 2022-12-02 #4840 @EtienneCmb
Hi @emdupre, I completed my review. This package has several originalities justifying its existence and its importance, especially for the neurophysiological community.  Among those originalities, the possibility to choose collapsing axes is an important feature for performing, for example, single-trial and dynamic estimations of the connectivity. As raised by @AJQuinn, the documentation can be improved. I still have two issues opened on GPU computations and comparison with other software.
