### 2022-12-02 e79302b
Fix name of environmental variable
Fix #53
