+++ spectral_connectivity/connectivity.py
@@ -19,7 +19,7 @@ from spectral_connectivity.statistics import (
     get_normal_distribution_p_values,
 )
 
-if os.environ.get("SPEC_CON_ENABLE_GPU") == "true":
+if os.environ.get("SPECTRAL_CONNECTIVITY_ENABLE_GPU") == "true":
     try:
         import cupy as xp
         from cupyx.scipy.fft import ifft
