### 2022-12-02 6b478fa
Add in simulation code
Re: issue #52
