### 2022-12-02 Eden-Kramer-Lab/spectral_connectivity#52 @edeno
Hi @AJQuinn , I believe I have addressed everything in this issue, so I will close it. Please let me know if I haven't addressed anything sufficiently. I am working on updating the documentation for the other issue currently.
