### 2022-12-02 Eden-Kramer-Lab/spectral_connectivity#50 @edeno
Hi @AJQuinn , I believe I have fixed all these issues in the documentation. Please reopen the issue if you feel that these haven't been fixed. I am working on the usage examples for the other issue.
