### 2022-12-02 61218c8
Add simulation module to docs
