+++ docs/api.rst
@@ -10,6 +10,7 @@ API
 
    spectral_connectivity.connectivity
    spectral_connectivity.minimum_phase_decomposition
+   spectral_connectivity.simulate
    spectral_connectivity.statistics
    spectral_connectivity.transforms
    spectral_connectivity.wrapper
