+++ docs/tutorials.rst
@@ -1,8 +1,0 @@
-Tutorials
-=========
-
-.. toctree::
-   :caption: Introductory tutorial
-   :maxdepth: 1
-
-   examples/Tutorial_On_Simulated_Examples.ipynb
