### 2022-12-02 a62c4cc
Add more helpful logging for whether GPU and CPU are being used.
