### 2022-12-02 40b7e99
Exclude setup.py and README
