+++ docs/examples/README.md
@@ -1,0 +1,1 @@
+Notebooks will be copied to this folder when building the documentation.
