+++ docs/index.md
@@ -145,6 +145,7 @@ that is rescued by antipsychotic drugs <https://doi.org/10.1101/2021.02.03.42958
 :hidden:
 :maxdepth: 2
 
+examples/Intro_tutorial
 examples/Tutorial_On_Simulated_Examples
 examples/Tutorial_Using_Paper_Examples
 
