### 2022-12-03 Eden-Kramer-Lab/spectral_connectivity#51 @edeno
Hi @AJQuinn , I've reworked the documentation and added a [tutorial](https://spectral-connectivity.readthedocs.io/en/latest/examples/Intro_tutorial.html). Please reopen if this doesn't match what you expected.
