### 2022-12-03 2492323
Add workaround for syntax highlighting in notebooks
