+++ spectral_connectivity/connectivity.py
@@ -101,7 +101,7 @@ class Connectivity:
     matrix.
 
     Note that spectral granger methods that require estimation of transfer function
-    and noise covariance use minimum phase decomposition [1] to decompose
+    and noise covariance use minimum phase decomposition [1]_ to decompose
     the cross spectral matrix into square roots, which then can be used to
     non-parametrically estimate the transfer function and noise covariance.
 
@@ -363,7 +363,7 @@ class Connectivity:
     @_asnumpy
     @_non_negative_frequencies(axis=-3)
     def imaginary_coherence(self):
-        """The normalized imaginary component of the cross-spectrum [1].
+        """The normalized imaginary component of the cross-spectrum [1]_.
 
         Projects the cross-spectrum onto the imaginary axis to mitigate the
         effect of volume-conducted dependencies. Assumes volume-conducted
@@ -391,7 +391,7 @@ class Connectivity:
         )
 
     def canonical_coherence(self, group_labels):
-        """Finds the maximal coherence between each combination of groups [1].
+        """Finds the maximal coherence between each combination of groups [1]_.
 
         The canonical coherence finds two sets of weights such that the
         coherence between the linear combination of group1 and the linear
@@ -461,7 +461,7 @@ class Connectivity:
 
     def global_coherence(self, max_rank=1):
         """The linear combinations of signals that capture the most coherent
-        power at each frequency and time window [1].
+        power at each frequency and time window [1]_.
 
         This is a frequency domain analog of PCA over signals at a given
         frequency/time window.
@@ -529,7 +529,7 @@ class Connectivity:
     @_non_negative_frequencies(axis=-3)
     def _phase_locking_value(self):
         """The cross-spectrum with the power for each signal scaled to
-        a magnitude of 1 [1].
+        a magnitude of 1 [1]_.
 
         The phase locking value attempts to mitigate power differences
         between realizations (tapers or trials) by treating all values of
@@ -562,7 +562,7 @@ class Connectivity:
     def phase_lag_index(self):
         """A non-parametric synchrony measure designed to mitigate power
         differences between realizations (tapers, trials) and
-        volume-conduction [1].
+        volume-conduction [1]_.
 
         The phase lag index is the average sign of the imaginary
         component of the cross-spectrum. The imaginary component sets
@@ -596,7 +596,7 @@ class Connectivity:
     @_non_negative_frequencies(-3)
     def weighted_phase_lag_index(self):
         """Weighted average of the phase lag index using the imaginary
-        coherency magnitudes as weights [1].
+        coherency magnitudes as weights [1]_.
 
         Returns
         -------
@@ -622,7 +622,7 @@ class Connectivity:
     @_asnumpy
     def debiased_squared_phase_lag_index(self):
         """The square of the phase lag index corrected for the positive
-        bias induced by using the magnitude of the complex cross-spectrum [1].
+        bias induced by using the magnitude of the complex cross-spectrum [1]_.
 
         Returns
         -------
@@ -647,7 +647,7 @@ class Connectivity:
     def debiased_squared_weighted_phase_lag_index(self):
         """The square of the weighted phase lag index corrected for the
         positive bias induced by using the magnitude of the complex
-        cross-spectrum [1].
+        cross-spectrum [1]_.
 
         Returns
         -------
@@ -690,7 +690,7 @@ class Connectivity:
     def pairwise_phase_consistency(self):
         """The square of the phase locking value corrected for the
         positive bias induced by using the magnitude of the complex
-        cross-spectrum [1].
+        cross-spectrum [1]_.
 
         Returns
         -------
@@ -714,7 +714,7 @@ class Connectivity:
     @_asnumpy
     def pairwise_spectral_granger_prediction(self):
[…]
