### 2022-12-03 9f53fc8
Revert "Fix reference formatting"
This reverts commit 101a97f8cd989827acbcda0b3991a37262f8fe91.
