### 2022-12-05 #4840 @EtienneCmb
Dear @emdupre, the last issue regarding performance claims has been addressed. From my side, this package checks all the boxes. I recommend this manuscript for publication in JOSS.
