### 2022-12-06 021b8fc
Eliminate redundant rst
