+++ docs/source/spectral_connectivity.rst
@@ -1,46 +1,0 @@
-spectral_connectivity package
-=============================
-
-Submodules
-----------
-
-spectral_connectivity.connectivity module
------------------------------------------
-
-.. automodule:: spectral_connectivity.connectivity
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-spectral_connectivity.minimum_phase_decomposition module
---------------------------------------------------------
-
-.. automodule:: spectral_connectivity.minimum_phase_decomposition
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-spectral_connectivity.statistics module
----------------------------------------
-
-.. automodule:: spectral_connectivity.statistics
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-spectral_connectivity.transforms module
----------------------------------------
-
-.. automodule:: spectral_connectivity.transforms
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: spectral_connectivity
-    :members:
-    :undoc-members:
-    :show-inheritance:
