+++ docs/source/modules.rst
@@ -1,7 +1,0 @@
-spectral_connectivity
-=====================
-
-.. toctree::
-   :maxdepth: 4
-
-   spectral_connectivity
