### 2022-12-06 9b33aa9
Try numpydoc option
