+++ docs/conf.py
@@ -39,6 +39,7 @@ extensions = [
 ]
 autosummary_generate = True
 add_module_names = False
+numpydoc_class_members_toctree = False
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
