### 2022-12-06 a2327f3
Fix docstring
