### 2022-12-06 #4840 @emdupre
Thank you @edeno for actioning these reviewer comments, and @EtienneCmb (and @sappelhoff , who I see I have not yet thanked !! 🙇 ) for confirming your completed reviews !

While we wait for additional comments from @AJQuinn , I'll perform a few editorial checks.
