### 2022-12-07 3861b77
Remove path insert
