+++ docs/conf.py
@@ -3,11 +3,7 @@
 
 import os
 import shutil
-import sys
 
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here.
-sys.path.insert(0, os.path.abspath("../"))
 
 # -- Project information -----------------------------------------------------
 
