### 2022-12-08 #4840 @AJQuinn
Hi all - @edeno et al have gone above and beyond in addressing my comments and I'm very happy to recommend spectral_connectivity for publication.

The documentation changes are excellent - I think this will be a massive benefit for future users. Thank you for your efforts.
