### 2022-12-08 #4840 @emdupre
Thank you for confirming, @AJQuinn !

@edeno , I've finished reviewing the software and paper, and I'm also very happy with the submission. I do have a few editorial requests on the paper itself :

- For the `CliffUnifyingPairwiseInteractions2022` reference, could you please update the bib entry as:
  ```
  @article{CliffUnifyingPairwiseInteractions2022,
    doi = {10.48550/ARXIV.2201.11941},
    url = {https://arxiv.org/abs/2201.11941},
    author = {Cliff, Oliver M. and Lizier, Joseph T. and Tsuchiya, Naotsugu and Fulcher, Ben D.},
    title = {Unifying Pairwise Interactions in Complex Dynamics},
    year = {2022},
    eprint={2007.01941},
    archivePrefix={arXiv},
    primaryClass={cs.LG}
  }
  ```
- For the `Rokem2020` reference, could you please update the bib entry as:
  ```
  @software{Rokem2020,
    author = {Rokem, Ariel and  Ivanov, Paul and Perez, Fernando and Trumpis, Mike and Gramfort, Alex },
    title = {Nitime: timeseries analysis for neuroscience data},
    url = {https://github.com/nipy/nitime},
    version = {0.9},
    date = {2020-12-19},
  }
  ```
  assuming you did intend to cite the 0.9 release, otherwise please amend as appropriate !
- For the `SchlaflyPythonpracticingneuroscientist2020` reference, could you please update the entry as:
  ```
  @article{SchlaflyPythonpracticingneuroscientist2020,
    doi = {10.22541/au.159363438.81020330},
    url = {https://doi.org/10.22541%2Fau.159363438.81020330},
    year = {2020},
    month = {jul},
    publisher = {Authorea, Inc.},
    author = {Emily Schlafly and Anthea Cheung and Samantha W Michalka and Paul A Lipton and Caroline Moore Kochlacs and Jason Bohland and Uri T Eden and Mark Kramer},
    title = {Python for the practicing~ neuroscientist: an online educational resource}
  }
  ```
  Following their "Cite as" recommendation ?
- Could you please remove all `file` fields from the paper.bib file
- Could you please capitalize the start of your footnote ; i.e., "For example" rather than "for example"
