+++ paper.bib
@@ -13,7 +13,6 @@
   doi = {10.1016/j.jneumeth.2014.11.012},
   isbn = {0165-0270},
   langid = {english},
-  file = {/Users/edeno/Dropbox (Personal)/Papers/Journal of Neuroscience Methods-/2015/Journal of Neuroscience Methods-2015-Aoi et al-Rate-adjusted spike–LFP coherence comparisons from spike-train statistics.pdf}
 }
 
 @inproceedings{BaccalaGeneralizedpartialdirected2007,
@@ -25,7 +24,6 @@
   pages = {163--166},
   publisher = {{IEEE}},
   doi = {10.1109/ICDSP.2007.4288544},
-  file = {/Users/edeno/Dropbox (Personal)/Papers/Digital Signal Processing, 2007 15th International Conference on-/2007/Digital Signal Processing, 2007 15th International Conference on-2007-Baccalá et al-Generalized partial directed coherence.pdf}
 }
 
 @article{BaccalaPartialdirectedcoherence2001,
@@ -43,7 +41,6 @@
   issn = {0340-1200},
   doi = {10.1007/PL00007990},
   langid = {english},
-  file = {/Users/edeno/Dropbox (Personal)/Papers/Biological Cybernetics-/2001/Biological Cybernetics-2001-Baccalá_Sameshima-Partial directed coherence.pdf}
 }
 
 @article{BaccalaStudyinginteractionbrain1998a,
@@ -68,18 +65,17 @@
   issn = {0027-8424, 1091-6490},
   doi = {10.1073/pnas.1017041108},
   langid = {english},
-  file = {/Users/edeno/Dropbox (Personal)/Papers/Proceedings of the National Academy of Sciences-/2011/Proceedings of the National Academy of Sciences-2011-Cimenser et al-Tracking brain states under general anesthesia by using global coherence.pdf}
 }
 
 @article{CliffUnifyingPairwiseInteractions2022,
-  title = {Unifying {{Pairwise Interactions}} in {{Complex Dynamics}}},
-  author = {Cliff, Oliver M and Lizier, Joseph T and Tsuchiya, Naotsugu and Fulcher, Ben D},
+  doi = {10.48550/ARXIV.2201.11941},
+  url = {https://arxiv.org/abs/2201.11941},
+  author = {Cliff, Oliver M. and Lizier, Joseph T. and Tsuchiya, Naotsugu and Fulcher, Ben D.},
+  title = {Unifying Pairwise Interactions in Complex Dynamics},
   year = {2022},
-  journal = {arXiv preprint arXiv:2201.11941},
-  eprint = {2201.11941},
-  eprinttype = {arxiv},
-  archiveprefix = {arXiv},
-  doi = {10.48550/arXiv.2201.11941}
+  eprint={2007.01941},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG}
 }
 
 @article{Delgado-SallentPhencyclidineinducedpsychosiscauses2021,
@@ -103,7 +99,6 @@
   issn = {10538119},
   doi = {10.1016/j.neuroimage.2008.02.020},
   langid = {english},
-  file = {/Users/edeno/Dropbox (Personal)/Papers/NeuroImage-/2008/NeuroImage-2008-Dhamala et al-Analyzing information flow in brain networks with nonparametric Granger.pdf}
 }
 
 @article{GewekeMeasurementLinearDependence1982,
@@ -119,7 +114,6 @@
   publisher = {{Taylor \& Francis}},
   issn = {01621459},
   doi = {10.2307/2287238},
[…]
