+++ spectral_connectivity/__init__.py
@@ -2,3 +2,5 @@
 from spectral_connectivity.connectivity import Connectivity
 from spectral_connectivity.transforms import Multitaper
 from spectral_connectivity.wrapper import multitaper_connectivity
+
+__version__ = "1.1.0"
