### 2022-12-09 #4840 @edeno
Hi @emdupre:

Thank you so much for your work and to all the reviewers. The process has been very educational.

I have:
- [x] Fixed bib entries
- [x] Removed all file fields from bib
- [x] Capitalized start of footnote
- [x] Made a tagged release, version tag: v1.1.0
- [x] Archived DOI: 10.5281/zenodo.7416614
