+++ paper.bib
@@ -1,4 +1,3 @@
-
 @article{AoiRateadjustedspikeLFP2015,
   ids = {AoiRateadjustedspikeLFP2015a},
   title = {Rate-Adjusted Spike\textendash{{LFP}} Coherence Comparisons from Spike-Train Statistics},
