### 2022-12-09 #4840 @emdupre
Thank you, @edeno !

I noticed that the metadata on the archive has several small discrepancies from the paper metadata. In particular, could you please update the title to :

> Spectral Connectivity: a python package for computing multitaper spectral estimates and frequency-domain brain connectivity measures on the CPU and GPU

and double-check that the author information lists names and surnames correctly ? From my rendering, these appear inverted in the final citation (i.e., excerpting from the bib file):

```
  author = {Eric, Denovellis and
                  Maxym, Myroshnychenko and
                  Mehrad, Sarmashghi and
                  Emily, Stephen},
```

You can update these without creating a new archive by editing the existing archive's metadata as described [here](https://help.zenodo.org/#technical).
