### 2022-12-09 #4840 @emdupre
Thank you ; this now looks right on my end !

I'm now happy to recommend Spectral Connectivity to the EIC team for publication, and I just want to add my congratulations to you on such an impressive effort ! 🎉
