### 2022-12-09 575c3f4
Fix typos
+ Added space between citation and last word.
+ Removed extra comma
+ Changed easily to ease
