### 2022-12-09 #4840 @edeno
I found a couple of small typos that I fixed. We are happy how it looks otherwise.
