### 2022-12-12 #4840 @Kevin-Mattheus-Moerman
@edeno I am the AEiC for this track and here to help process this work for acceptance in JOSS. I have reviewed the archive and the paper and all seems in order. I only have one minor editorial point:

- [ ] Please add the affiliation country to all affiliations, and do not use acronyms, i.e use `United States of America` instead of `USA`.

Please let me know when you've worked on the above. Thanks.
