+++ CITATION.cff
@@ -6,7 +6,7 @@ authors:
   - family-names: Denovellis
     given-names: Eric
     orcid: https://orcid.org/0000-0003-4606-087X
-    affiliation: Howard Hughes Medical Institute - University of California, San Francisco, California, United States of America
+    affiliation: Howard Hughes Medical Institute - University of California, San Francisco, San Francisco, California, United States of America
   - family-names: Myroshnychenko
     given-names: Maxym
     affiliation: National Institute of Neurological Disorders and Stroke, Bethesda, Maryland, United States of America
