### 2022-12-13 #4840 @Kevin-Mattheus-Moerman
@edeno congratulations on this JOSS publication! :tada:

Thanks @emdupre for editing!

And also thank you to the reviewers @AJQuinn, @sappelhoff, and @EtienneCmb!!!!!
